+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.7.2 Under development
 
-- Enh: Move `normalizeKey()` to `AttributeBag` and add optional `$prefix` support to `get()`, `remove()`, `set()`, and `setMany()` (@terabytesoftw)
+- Enh #52: Move `normalizeKey()` to `AttributeBag` and add optional `$prefix` support to `get()`, `remove()`, `set()`, and `setMany()` (@terabytesoftw)
 
 ## 0.7.1 February 15, 2026